Add tests for webpack update task
