+++ flow-server/src/test/java/com/vaadin/flow/server/frontend/NodeUpdatePackagesTest.java
@@ -19,10 +19,6 @@ package com.vaadin.flow.server.frontend;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.util.List;
-import java.util.Objects;
-import java.util.stream.Collectors;
 
 import org.junit.Assert;
 import org.junit.Before;
@@ -33,8 +29,6 @@ import org.junit.rules.TemporaryFolder;
 import elemental.json.JsonObject;
 
 import static com.vaadin.flow.server.Constants.PACKAGE_JSON;
-import static com.vaadin.flow.server.frontend.FrontendUtils.FLOW_IMPORTS_FILE;
-import static com.vaadin.flow.server.frontend.FrontendUtils.WEBPACK_CONFIG;
 import static com.vaadin.flow.server.frontend.FrontendUtils.getBaseDir;
 
 public class NodeUpdatePackagesTest extends NodeUpdateTestUtil {
@@ -43,10 +37,8 @@ public class NodeUpdatePackagesTest extends NodeUpdateTestUtil {
     public TemporaryFolder temporaryFolder = new TemporaryFolder();
 
     private TaskUpdatePackages packageUpdater;
-    private TaskUpdateWebpack webpackUpdater;
     private TaskCreatePackageJson packageCreator;
     private File packageJson;
-    private File webpackConfig;
 
     @Before
     public void setup() throws Exception {
@@ -62,13 +54,7 @@ public class NodeUpdatePackagesTest extends NodeUpdateTestUtil {
         packageUpdater = new TaskUpdatePackages(getClassFinder(), baseDir,
                 new File(baseDir, "node_modules"), true);
 
-
-
-        webpackUpdater = new TaskUpdateWebpack(baseDir, baseDir, WEBPACK_CONFIG,
-                new File(baseDir, FLOW_IMPORTS_FILE));
-
         packageJson = new File(baseDir, PACKAGE_JSON);
-        webpackConfig = new File(baseDir, WEBPACK_CONFIG);
     }
 
     @Test
@@ -78,12 +64,6 @@ public class NodeUpdatePackagesTest extends NodeUpdateTestUtil {
         Assert.assertTrue(packageJson.exists());
     }
 
-    @Test
-    public void should_CreateWebpackConfig() throws Exception {
-        Assert.assertFalse(webpackConfig.exists());
-        webpackUpdater.execute();
-        assertWebpackConfigContent();
-    }
 
     @Test
     public void should_not_ModifyPackageJson_WhenAlreadyExists() throws Exception {
@@ -132,44 +112,4 @@ public class NodeUpdatePackagesTest extends NodeUpdateTestUtil {
                 devDependencies.hasKey("copy-webpack-plugin"));
     }
 
-    private void assertWebpackConfigContent() throws IOException {
-        List<String> webpackContents = Files.lines(webpackConfig.toPath()).collect(Collectors.toList());
-
-        Assert.assertFalse(
-                "webpack config should not contain Windows path separators",
-                webpackContents.contains("\\\\"));
-
-        verifyNoAbsolutePathsPresent(webpackContents);
-    }
-
-    private void verifyNoAbsolutePathsPresent(List<String> webpackContents) {
-        List<String> wrongLines = webpackContents.stream()
-            // check the lines with slashes only
-            .filter(line -> line.contains("/"))
-            // trim the whitespaces
-            .map(line -> line.replaceAll("\\s", ""))
-            // check the equals ( a=something ) and object declarations ( {a: something} )
-            .map(line -> {
-                    int equalsSignPosition = line.indexOf("=");
-                    int jsonPropertySignPosition = line.indexOf(":");
-                    if (equalsSignPosition > 0) {
-                        return line.substring(equalsSignPosition + 1);
-                    } else if (jsonPropertySignPosition > 0) {
-                        return line.substring(jsonPropertySignPosition + 1);
-                    } else {
-                        return null;
-                    }
-                })
-            .filter(Objects::nonNull)
-            // take the lines with strings only and trim the string start
-            .filter(line -> line.startsWith("'") || line.startsWith("\"") || line.startsWith("`"))
-            .map(line -> line.substring(1))
-            .filter(line -> line.startsWith("/"))
-            .collect(Collectors.toList());
-
-        Assert.assertTrue(String.format(
-                "Expected to have no lines that have a string starting with a slash in assignment. Incorrect lines: '%s'",
-                wrongLines),
-                wrongLines.isEmpty());
-    }
 }
